pyproject.toml
[tool.poetry]
name = "spectrumone_test"
version = "0.1.0"
description = ""
authors = ["Joween Flores <4754752+joweenflores@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.4"}
python-dotenv = "^0.20.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
